+++ include/softmax_layer.h
@@ -7,5 +7,5 @@ class softmax_layer : public fc_layer
 	void compute_derivatives(const fc_layer& next) override;
 
 public:
-	explicit softmax_layer(size_t prev_size, size_t size);
+	explicit softmax_layer(int prev_size, int size);
 };
